hip new user interface
Colors are temporary, I will probably make a texture for the borders.
Font is supposed to be treasure map deadhand, but it's not working (it's
in the content/resource/ folder). Team change menu fixed, but you need
to use the console command (show_team_menu) to access it. Ship health
isn't working yet but the round counter should be.